Refactor: move NodeManager to standalone file.
