+++ example-raft-kv/src/network/mod.rs
@@ -1,4 +1,5 @@
 pub mod api;
 pub mod management;
+pub mod node_manager;
 pub mod raft;
-pub mod rpc;
+pub mod raft_network_impl;
